ViewModel, UseCaseを追加
